pr 피드백 수정 * 타입 부분 수정
