+++ coffee_fe/interfaces/rq/coffeeRq.ts
@@ -2,10 +2,3 @@ export interface CoffeeRq {
     name: string;
 }
 
-export interface OrderRq {
-    orderID: number;
-    orderDate: string;
-    orderStatus: string;
-    menuId: number;
-    orderMenu: string;
-}
