+++ coffee_fe/interfaces/rq/orderRq.ts
@@ -1,0 +1,7 @@
+export interface OrderRq {
+    orderID: number;
+    orderDate: string;
+    orderStatus: string;
+    menuId: number;
+    orderMenu: string;
+}
