+++ coffee_fe/interfaces/rs/coffeeRs.ts
@@ -3,12 +3,3 @@ export interface CoffeeRs {
     menuName: string;
     category: string;
 }
-
-export interface OrderRs {
-    orderID: number;
-    orderDate: string;
-    orderStatus: string;
-    menuId: number;
-    orderMenu: string;
-}
-
