+++ coffee_fe/interfaces/rs/orderRs.ts
@@ -1,0 +1,10 @@
+import {CoffeeRs} from "./coffeeRs";
+
+export interface OrderRs {
+    orderID: number;
+    orderDate: string;
+    orderStatus: string;
+    menuId: CoffeeRs[];
+    orderMenu: string;
+}
+
